+++ adapter/src/components/ServerVersionProvider.js
@@ -23,6 +23,7 @@ export const ServerVersionProvider = ({
     plugin,
     parentAlertsAdd,
     showAlertsInPlugin,
+    loginApp,
     children,
 }) => {
     const offlineInterface = useOfflineInterface()
@@ -109,6 +110,29 @@ export const ServerVersionProvider = ({
                 ? state
                 : { loading: true, error: undefined, systemInfo: undefined }
         )
+
+        // version is available from api/loginConfig
+        if (loginApp) {
+            const requestLogin = get(`${baseUrl}/api/loginConfig`)
+            requestLogin
+                .then((loginConfig) => {
+                    setSystemInfoState({
+                        loading: false,
+                        error: undefined,
+                        systemInfo: {
+                            version: loginConfig.apiVersion ?? '2.41',
+                        },
+                    })
+                })
+                .catch((e) => {
+                    setSystemInfoState({
+                        loading: false,
+                        error: e,
+                        systemInfo: undefined,
+                    })
+                })
+            return
+        }
         const request = get(`${baseUrl}/api/system/info`)
         request
             .then((systemInfo) => {
@@ -131,7 +155,7 @@ export const ServerVersionProvider = ({
         return () => {
             request.abort()
         }
-    }, [appName, baseUrl])
+    }, [appName, baseUrl, loginApp])
 
     useEffect(() => {
         if (pwaEnabled) {
@@ -143,7 +167,13 @@ export const ServerVersionProvider = ({
 
     // This needs to come before 'loading' case to show modal at correct times
     if (systemInfoState.error || baseUrlState.error) {
-        return <LoginModal appName={appName} baseUrl={baseUrl} />
+        return (
+            <LoginModal
+                appName={appName}
+                baseUrl={baseUrl}
+                loginApp={loginApp}
+            />
+        )
     }
 
     if (
@@ -171,7 +201,7 @@ export const ServerVersionProvider = ({
                 systemInfo,
                 pwaEnabled,
             }}
-            offlineInterface={offlineInterface}
+            offlineInterface={loginApp ? null : offlineInterface}
             plugin={plugin}
             parentAlertsAdd={parentAlertsAdd}
             showAlertsInPlugin={showAlertsInPlugin}
@@ -186,6 +216,7 @@ ServerVersionProvider.propTypes = {
     appVersion: PropTypes.string.isRequired,
     apiVersion: PropTypes.number,
     children: PropTypes.element,
+    loginApp: PropTypes.bool,
     parentAlertsAdd: PropTypes.func,
     plugin: PropTypes.bool,
     pwaEnabled: PropTypes.bool,