+++ cli/config/d2ConfigDefaults.typescript.js
@@ -1,0 +1,17 @@
+const defaultsApp = {
+    type: 'app',
+
+    entryPoints: {
+        app: './src/App.tsx',
+    },
+}
+
+const defaultsLib = {
+    type: 'lib',
+
+    entryPoints: {
+        lib: './src/index.tsx',
+    },
+}
+
+module.exports = { defaultsApp, defaultsLib }