+++ cli/src/lib/plugin/index.js
@@ -1,7 +1,0 @@
-const build = require('./build')
-const start = require('./start')
-
-module.exports = ({ config, paths, pluginifiedApp = false }) => ({
-    build: () => build({ config, paths, pluginifiedApp }),
-    start: ({ port }) => start({ port, config, paths, pluginifiedApp }),
-})